requirements.txt
python-fasthtml
markupsafe